lint: rename partitioners
